% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\docType{package}
\name{bigquery_googleAuthR}
\alias{bigquery_googleAuthR}
\alias{bigquery_googleAuthR-package}
\title{BigQuery API
A data platform for customers to create, manage, share and query data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:28:52
filename: /Users/mark/dev/R/autoGoogleAPI/googlebigqueryv2.auto/R/bigquery_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/bigquery.insertdata
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}
}
