% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{BigtableColumn}
\alias{BigtableColumn}
\title{BigQuery API Objects 
A data platform for customers to create, manage, share and query data.}
\usage{
BigtableColumn(encoding = NULL, fieldName = NULL, onlyReadLatest = NULL,
  qualifierEncoded = NULL, qualifierString = NULL, type = NULL)
}
\arguments{
\item{encoding}{[Optional] The encoding of the values when the type is not STRING}

\item{fieldName}{[Optional] If the qualifier is not a valid BigQuery field identifier i}

\item{onlyReadLatest}{[Optional] If this is set, only the latest version of value in this column are exposed}

\item{qualifierEncoded}{[Required] Qualifier of the column}

\item{qualifierString}{No description}

\item{type}{[Optional] The type to convert the value in cells of this column}
}
\value{
BigtableColumn object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:57:35
filename: /Users/mark/dev/R/autoGoogleAPI/googlebigqueryv2.auto/R/bigquery_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BigtableColumn Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

