% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{BigtableColumnFamily}
\alias{BigtableColumnFamily}
\title{BigtableColumnFamily Object}
\usage{
BigtableColumnFamily(columns = NULL, encoding = NULL, familyId = NULL,
  onlyReadLatest = NULL, type = NULL)
}
\arguments{
\item{columns}{[Optional] Lists of columns that should be exposed as individual fields as opposed to a list of (column name, value) pairs}

\item{encoding}{[Optional] The encoding of the values when the type is not STRING}

\item{familyId}{Identifier of the column family}

\item{onlyReadLatest}{[Optional] If this is set only the latest version of value are exposed for all columns in this column family}

\item{type}{[Optional] The type to convert the value in cells of this column family}
}
\value{
BigtableColumnFamily object
}
\description{
BigtableColumnFamily Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

