% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{CsvOptions}
\alias{CsvOptions}
\title{CsvOptions Object}
\usage{
CsvOptions(allowJaggedRows = NULL, allowQuotedNewlines = NULL,
  encoding = NULL, fieldDelimiter = NULL, quote = NULL,
  skipLeadingRows = NULL)
}
\arguments{
\item{allowJaggedRows}{[Optional] Indicates if BigQuery should accept rows that are missing trailing optional columns}

\item{allowQuotedNewlines}{[Optional] Indicates if BigQuery should allow quoted data sections that contain newline characters in a CSV file}

\item{encoding}{[Optional] The character encoding of the data}

\item{fieldDelimiter}{[Optional] The separator for fields in a CSV file}

\item{quote}{[Optional] The value that is used to quote data sections in a CSV file}

\item{skipLeadingRows}{[Optional] The number of rows at the top of a CSV file that BigQuery will skip when reading the data}
}
\value{
CsvOptions object
}
\description{
CsvOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
