% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Object}
\usage{
Dataset(Dataset.access = NULL, Dataset.labels = NULL, access = NULL,
  creationTime = NULL, datasetReference = NULL,
  defaultTableExpirationMs = NULL, description = NULL, etag = NULL,
  friendlyName = NULL, id = NULL, labels = NULL,
  lastModifiedTime = NULL, location = NULL, selfLink = NULL)
}
\arguments{
\item{Dataset.access}{The \link{Dataset.access} object or list of objects}

\item{Dataset.labels}{The \link{Dataset.labels} object or list of objects}

\item{access}{[Optional] An array of objects that define dataset access for one or more entities}

\item{creationTime}{[Output-only] The time when this dataset was created, in milliseconds since the epoch}

\item{datasetReference}{[Required] A reference that identifies the dataset}

\item{defaultTableExpirationMs}{[Optional] The default lifetime of all tables in the dataset, in milliseconds}

\item{description}{[Optional] A user-friendly description of the dataset}

\item{etag}{[Output-only] A hash of the resource}

\item{friendlyName}{[Optional] A descriptive name for the dataset}

\item{id}{[Output-only] The fully-qualified unique name of the dataset in the format projectId:datasetId}

\item{labels}{[Experimental] The labels associated with this dataset}

\item{lastModifiedTime}{[Output-only] The date when this dataset or any of its tables was last modified, in milliseconds since the epoch}

\item{location}{[Experimental] The geographic location where the dataset should reside}

\item{selfLink}{[Output-only] A URL that can be used to access the resource again}
}
\value{
Dataset object
}
\description{
Dataset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Dataset functions: \code{\link{Dataset.access}},
  \code{\link{Dataset.labels}},
  \code{\link{datasets.insert}},
  \code{\link{datasets.patch}},
  \code{\link{datasets.update}}
}

