% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{DatasetList.datasets}
\alias{DatasetList.datasets}
\title{DatasetList.datasets Object}
\usage{
DatasetList.datasets(DatasetList.datasets.labels = NULL)
}
\arguments{
\item{DatasetList.datasets.labels}{The \link{DatasetList.datasets.labels} object or list of objects}
}
\value{
DatasetList.datasets object
}
\description{
DatasetList.datasets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DatasetList functions: \code{\link{DatasetList.datasets.labels}},
  \code{\link{DatasetList}}
}
