% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{DatasetList}
\alias{DatasetList}
\title{DatasetList Object}
\usage{
DatasetList(DatasetList.datasets = NULL, DatasetList.datasets.labels = NULL,
  datasets = NULL, etag = NULL, nextPageToken = NULL)
}
\arguments{
\item{DatasetList.datasets}{The \link{DatasetList.datasets} object or list of objects}

\item{DatasetList.datasets.labels}{The \link{DatasetList.datasets.labels} object or list of objects}

\item{datasets}{An array of the dataset resources in the project}

\item{etag}{A hash value of the results page}

\item{nextPageToken}{A token that can be used to request the next results page}
}
\value{
DatasetList object
}
\description{
DatasetList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DatasetList functions: \code{\link{DatasetList.datasets.labels}},
  \code{\link{DatasetList.datasets}}
}
