% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{DatasetReference}
\alias{DatasetReference}
\title{DatasetReference Object}
\usage{
DatasetReference(datasetId = NULL, projectId = NULL)
}
\arguments{
\item{datasetId}{[Required] A unique ID for this dataset, without the project name}

\item{projectId}{[Optional] The ID of the project containing this dataset}
}
\value{
DatasetReference object
}
\description{
DatasetReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
