% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{datasets.list}
\alias{datasets.list}
\title{Lists all datasets in the specified project to which you have been granted the READER dataset role.}
\usage{
datasets.list(projectId, all = NULL, filter = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{projectId}{Project ID of the datasets to be listed}

\item{all}{Whether to list all datasets, including hidden ones}

\item{filter}{An expression for filtering the results of the request by label}

\item{maxResults}{The maximum number of results to return}

\item{pageToken}{Page token, returned by a previous call, to request the next page of results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}

