% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{datasets.patch}
\alias{datasets.patch}
\title{Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource. This method supports patch semantics.}
\usage{
datasets.patch(Dataset, projectId, datasetId)
}
\arguments{
\item{Dataset}{The \link{Dataset} object to pass to this method}

\item{projectId}{Project ID of the dataset being updated}

\item{datasetId}{Dataset ID of the dataset being updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}

Other Dataset functions: \code{\link{Dataset.access}},
  \code{\link{Dataset.labels}}, \code{\link{Dataset}},
  \code{\link{datasets.insert}},
  \code{\link{datasets.update}}
}

