% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ErrorProto}
\alias{ErrorProto}
\title{ErrorProto Object}
\usage{
ErrorProto(debugInfo = NULL, location = NULL, message = NULL,
  reason = NULL)
}
\arguments{
\item{debugInfo}{Debugging information}

\item{location}{Specifies where the error occurred, if present}

\item{message}{A human-readable description of the error}

\item{reason}{A short error code that summarizes the error}
}
\value{
ErrorProto object
}
\description{
ErrorProto Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
