% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ExplainQueryStage}
\alias{ExplainQueryStage}
\title{ExplainQueryStage Object}
\usage{
ExplainQueryStage(computeRatioAvg = NULL, computeRatioMax = NULL,
  id = NULL, name = NULL, readRatioAvg = NULL, readRatioMax = NULL,
  recordsRead = NULL, recordsWritten = NULL, steps = NULL,
  waitRatioAvg = NULL, waitRatioMax = NULL, writeRatioAvg = NULL,
  writeRatioMax = NULL)
}
\arguments{
\item{computeRatioAvg}{Relative amount of time the average shard spent on CPU-bound tasks}

\item{computeRatioMax}{Relative amount of time the slowest shard spent on CPU-bound tasks}

\item{id}{Unique ID for stage within plan}

\item{name}{Human-readable name for stage}

\item{readRatioAvg}{Relative amount of time the average shard spent reading input}

\item{readRatioMax}{Relative amount of time the slowest shard spent reading input}

\item{recordsRead}{Number of records read into the stage}

\item{recordsWritten}{Number of records written by the stage}

\item{steps}{List of operations within the stage in dependency order (approximately chronological)}

\item{waitRatioAvg}{Relative amount of time the average shard spent waiting to be scheduled}

\item{waitRatioMax}{Relative amount of time the slowest shard spent waiting to be scheduled}

\item{writeRatioAvg}{Relative amount of time the average shard spent on writing output}

\item{writeRatioMax}{Relative amount of time the slowest shard spent on writing output}
}
\value{
ExplainQueryStage object
}
\description{
ExplainQueryStage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

