% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ExplainQueryStep}
\alias{ExplainQueryStep}
\title{ExplainQueryStep Object}
\usage{
ExplainQueryStep(substeps = NULL)
}
\arguments{
\item{substeps}{Human-readable stage descriptions}
}
\value{
ExplainQueryStep object
}
\description{
ExplainQueryStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

