% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ExternalDataConfiguration}
\alias{ExternalDataConfiguration}
\title{ExternalDataConfiguration Object}
\usage{
ExternalDataConfiguration(autodetect = NULL, bigtableOptions = NULL,
  compression = NULL, csvOptions = NULL, googleSheetsOptions = NULL,
  ignoreUnknownValues = NULL, maxBadRecords = NULL, schema = NULL,
  sourceFormat = NULL, sourceUris = NULL)
}
\arguments{
\item{autodetect}{[Experimental] Try to detect schema and format options automatically}

\item{bigtableOptions}{[Optional] Additional options if sourceFormat is set to BIGTABLE}

\item{compression}{[Optional] The compression type of the data source}

\item{csvOptions}{Additional properties to set if sourceFormat is set to CSV}

\item{googleSheetsOptions}{[Optional] Additional options if sourceFormat is set to GOOGLE_SHEETS}

\item{ignoreUnknownValues}{[Optional] Indicates if BigQuery should allow extra values that are not represented in the table schema}

\item{maxBadRecords}{[Optional] The maximum number of bad records that BigQuery can ignore when reading data}

\item{schema}{[Optional] The schema for the data}

\item{sourceFormat}{[Required] The data format}

\item{sourceUris}{[Required] The fully-qualified URIs that point to your data in Google Cloud}
}
\value{
ExternalDataConfiguration object
}
\description{
ExternalDataConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

