% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{GoogleSheetsOptions}
\alias{GoogleSheetsOptions}
\title{GoogleSheetsOptions Object}
\usage{
GoogleSheetsOptions(skipLeadingRows = NULL)
}
\arguments{
\item{skipLeadingRows}{[Optional] The number of rows at the top of a sheet that BigQuery will skip when reading the data}
}
\value{
GoogleSheetsOptions object
}
\description{
GoogleSheetsOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
