% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{Job}
\alias{Job}
\title{Job Object}
\usage{
Job(configuration = NULL, etag = NULL, id = NULL, jobReference = NULL,
  selfLink = NULL, statistics = NULL, status = NULL, user_email = NULL)
}
\arguments{
\item{configuration}{[Required] Describes the job configuration}

\item{etag}{[Output-only] A hash of this resource}

\item{id}{[Output-only] Opaque ID field of the job}

\item{jobReference}{[Optional] Reference describing the unique-per-user name of the job}

\item{selfLink}{[Output-only] A URL that can be used to access this resource again}

\item{statistics}{[Output-only] Information about the job, including starting time and ending time of the job}

\item{status}{[Output-only] The status of this job}

\item{user_email}{[Output-only] Email address of the user who ran the job}
}
\value{
Job object
}
\description{
Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Job functions: \code{\link{jobs.insert}}
}
