% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobCancelResponse}
\alias{JobCancelResponse}
\title{JobCancelResponse Object}
\usage{
JobCancelResponse(job = NULL)
}
\arguments{
\item{job}{The final state of the job}
}
\value{
JobCancelResponse object
}
\description{
JobCancelResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
