% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobConfigurationLoad}
\alias{JobConfigurationLoad}
\title{JobConfigurationLoad Object}
\usage{
JobConfigurationLoad(allowJaggedRows = NULL, allowQuotedNewlines = NULL,
  autodetect = NULL, createDisposition = NULL, destinationTable = NULL,
  encoding = NULL, fieldDelimiter = NULL, ignoreUnknownValues = NULL,
  maxBadRecords = NULL, nullMarker = NULL, projectionFields = NULL,
  quote = NULL, schema = NULL, schemaInline = NULL,
  schemaInlineFormat = NULL, schemaUpdateOptions = NULL,
  skipLeadingRows = NULL, sourceFormat = NULL, sourceUris = NULL,
  writeDisposition = NULL)
}
\arguments{
\item{allowJaggedRows}{[Optional] Accept rows that are missing trailing optional columns}

\item{allowQuotedNewlines}{Indicates if BigQuery should allow quoted data sections that contain newline characters in a CSV file}

\item{autodetect}{[Experimental] Indicates if we should automatically infer the options and schema for CSV and JSON sources}

\item{createDisposition}{[Optional] Specifies whether the job is allowed to create new tables}

\item{destinationTable}{[Required] The destination table to load the data into}

\item{encoding}{[Optional] The character encoding of the data}

\item{fieldDelimiter}{[Optional] The separator for fields in a CSV file}

\item{ignoreUnknownValues}{[Optional] Indicates if BigQuery should allow extra values that are not represented in the table schema}

\item{maxBadRecords}{[Optional] The maximum number of bad records that BigQuery can ignore when running the job}

\item{nullMarker}{[Optional] Specifies a string that represents a null value in a CSV file}

\item{projectionFields}{[Experimental] If sourceFormat is set to 'DATASTORE_BACKUP', indicates which entity properties to load into BigQuery from a Cloud Datastore backup}

\item{quote}{[Optional] The value that is used to quote data sections in a CSV file}

\item{schema}{[Optional] The schema for the destination table}

\item{schemaInline}{[Deprecated] The inline schema}

\item{schemaInlineFormat}{[Deprecated] The format of the schemaInline property}

\item{schemaUpdateOptions}{[Experimental] Allows the schema of the desitination table to be updated as a side effect of the load job}

\item{skipLeadingRows}{[Optional] The number of rows at the top of a CSV file that BigQuery will skip when loading the data}

\item{sourceFormat}{[Optional] The format of the data files}

\item{sourceUris}{[Required] The fully-qualified URIs that point to your data in Google Cloud Storage}

\item{writeDisposition}{[Optional] Specifies the action that occurs if the destination table already exists}
}
\value{
JobConfigurationLoad object
}
\description{
JobConfigurationLoad Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
