% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobConfigurationQuery}
\alias{JobConfigurationQuery}
\title{JobConfigurationQuery Object}
\usage{
JobConfigurationQuery(JobConfigurationQuery.tableDefinitions = NULL,
  allowLargeResults = NULL, createDisposition = NULL,
  defaultDataset = NULL, destinationTable = NULL, flattenResults = NULL,
  maximumBillingTier = NULL, maximumBytesBilled = NULL,
  parameterMode = NULL, preserveNulls = NULL, priority = NULL,
  query = NULL, queryParameters = NULL, schemaUpdateOptions = NULL,
  tableDefinitions = NULL, useLegacySql = NULL, useQueryCache = NULL,
  userDefinedFunctionResources = NULL, writeDisposition = NULL)
}
\arguments{
\item{JobConfigurationQuery.tableDefinitions}{The \link{JobConfigurationQuery.tableDefinitions} object or list of objects}

\item{allowLargeResults}{If true, allows the query to produce arbitrarily large result tables at a slight cost in performance}

\item{createDisposition}{[Optional] Specifies whether the job is allowed to create new tables}

\item{defaultDataset}{[Optional] Specifies the default dataset to use for unqualified table names in the query}

\item{destinationTable}{[Optional] Describes the table where the query results should be stored}

\item{flattenResults}{[Optional] Flattens all nested and repeated fields in the query results}

\item{maximumBillingTier}{[Optional] Limits the billing tier for this job}

\item{maximumBytesBilled}{[Optional] Limits the bytes billed for this job}

\item{parameterMode}{[Experimental] Standard SQL only}

\item{preserveNulls}{[Deprecated] This property is deprecated}

\item{priority}{[Optional] Specifies a priority for the query}

\item{query}{[Required] BigQuery SQL query to execute}

\item{queryParameters}{Query parameters for standard SQL queries}

\item{schemaUpdateOptions}{[Experimental] Allows the schema of the destination table to be updated as a side effect of the query job}

\item{tableDefinitions}{[Optional] If querying an external data source outside of BigQuery, describes the data format, location and other properties of the data source}

\item{useLegacySql}{Specifies whether to use BigQuery's legacy SQL dialect for this query}

\item{useQueryCache}{[Optional] Whether to look for the result in the query cache}

\item{userDefinedFunctionResources}{[Experimental] Describes user-defined function resources used in the query}

\item{writeDisposition}{[Optional] Specifies the action that occurs if the destination table already exists}
}
\value{
JobConfigurationQuery object
}
\description{
JobConfigurationQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other JobConfigurationQuery functions: \code{\link{JobConfigurationQuery.tableDefinitions}}
}
