% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobConfigurationQuery.tableDefinitions}
\alias{JobConfigurationQuery.tableDefinitions}
\title{JobConfigurationQuery.tableDefinitions Object}
\usage{
JobConfigurationQuery.tableDefinitions()
}
\value{
JobConfigurationQuery.tableDefinitions object
}
\description{
JobConfigurationQuery.tableDefinitions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] If querying an external data source outside of BigQuery, describes the data format, location and other properties of the data source. By defining these properties, the data source can then be queried as if it were a standard BigQuery table.
}
\seealso{
Other JobConfigurationQuery functions: \code{\link{JobConfigurationQuery}}
}

