% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobList.jobs}
\alias{JobList.jobs}
\title{JobList.jobs Object}
\usage{
JobList.jobs()
}
\value{
JobList.jobs object
}
\description{
JobList.jobs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other JobList functions: \code{\link{JobList}}
}
