% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobList}
\alias{JobList}
\title{JobList Object}
\usage{
JobList(JobList.jobs = NULL, etag = NULL, jobs = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{JobList.jobs}{The \link{JobList.jobs} object or list of objects}

\item{etag}{A hash of this page of results}

\item{jobs}{List of jobs that were requested}

\item{nextPageToken}{A token to request the next page of results}
}
\value{
JobList object
}
\description{
JobList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other JobList functions: \code{\link{JobList.jobs}}
}
