% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobReference}
\alias{JobReference}
\title{JobReference Object}
\usage{
JobReference(jobId = NULL, projectId = NULL)
}
\arguments{
\item{jobId}{[Required] The ID of the job}

\item{projectId}{[Required] The ID of the project containing this job}
}
\value{
JobReference object
}
\description{
JobReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
