% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{jobs.cancel}
\alias{jobs.cancel}
\title{Requests that a job be cancelled. This call will return immediately, and the client will need to poll for the job status to see if the cancel completed successfully. Cancelled jobs may still incur costs.}
\usage{
jobs.cancel(projectId, jobId)
}
\arguments{
\item{projectId}{[Required] Project ID of the job to cancel}

\item{jobId}{[Required] Job ID of the job to cancel}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}
