% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{jobs.getQueryResults}
\alias{jobs.getQueryResults}
\title{Retrieves the results of a query job.}
\usage{
jobs.getQueryResults(projectId, jobId, maxResults = NULL, pageToken = NULL,
  startIndex = NULL, timeoutMs = NULL)
}
\arguments{
\item{projectId}{[Required] Project ID of the query job}

\item{jobId}{[Required] Job ID of the query job}

\item{maxResults}{Maximum number of results to read}

\item{pageToken}{Page token, returned by a previous call, to request the next page of results}

\item{startIndex}{Zero-based index of the starting row}

\item{timeoutMs}{How long to wait for the query to complete, in milliseconds, before returning}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}
