% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{jobs.insert}
\alias{jobs.insert}
\title{Starts a new asynchronous job. Requires the Can View project role.}
\usage{
jobs.insert(Job, projectId)
}
\arguments{
\item{Job}{The \link{Job} object to pass to this method}

\item{projectId}{Project ID of the project that will be billed for the job}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_only, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}

Other Job functions: \code{\link{Job}}
}
