% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{jobs.list}
\alias{jobs.list}
\title{Lists all jobs that you started in the specified project. Job information is available for a six month period after creation. The job list is sorted in reverse chronological order, by job creation time. Requires the Can View project role, or the Is Owner project role if you set the allUsers property.}
\usage{
jobs.list(projectId, allUsers = NULL, maxResults = NULL, pageToken = NULL,
  projection = NULL, stateFilter = NULL)
}
\arguments{
\item{projectId}{Project ID of the jobs to list}

\item{allUsers}{Whether to display jobs owned by all users in the project}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Page token, returned by a previous call, to request the next page of results}

\item{projection}{Restrict information returned to a set of selected fields}

\item{stateFilter}{Filter for job state}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}

