% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{jobs.query}
\alias{jobs.query}
\title{Runs a BigQuery SQL query synchronously and returns query results if the query completes within a specified timeout.}
\usage{
jobs.query(QueryRequest, projectId)
}
\arguments{
\item{QueryRequest}{The \link{QueryRequest} object to pass to this method}

\item{projectId}{Project ID of the project billed for the query}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}

Other QueryRequest functions: \code{\link{QueryRequest}}
}
