% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobStatistics}
\alias{JobStatistics}
\title{JobStatistics Object}
\usage{
JobStatistics(creationTime = NULL, endTime = NULL, extract = NULL,
  load = NULL, query = NULL, startTime = NULL,
  totalBytesProcessed = NULL)
}
\arguments{
\item{creationTime}{[Output-only] Creation time of this job, in milliseconds since the epoch}

\item{endTime}{[Output-only] End time of this job, in milliseconds since the epoch}

\item{extract}{[Output-only] Statistics for an extract job}

\item{load}{[Output-only] Statistics for a load job}

\item{query}{[Output-only] Statistics for a query job}

\item{startTime}{[Output-only] Start time of this job, in milliseconds since the epoch}

\item{totalBytesProcessed}{[Output-only] [Deprecated] Use the bytes processed in the query statistics instead}
}
\value{
JobStatistics object
}
\description{
JobStatistics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
