% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobStatistics2}
\alias{JobStatistics2}
\title{JobStatistics2 Object}
\usage{
JobStatistics2(billingTier = NULL, cacheHit = NULL,
  numDmlAffectedRows = NULL, queryPlan = NULL, referencedTables = NULL,
  schema = NULL, statementType = NULL, totalBytesBilled = NULL,
  totalBytesProcessed = NULL, undeclaredQueryParameters = NULL)
}
\arguments{
\item{billingTier}{[Output-only] Billing tier for the job}

\item{cacheHit}{[Output-only] Whether the query result was fetched from the query cache}

\item{numDmlAffectedRows}{[Output-only, Experimental] The number of rows affected by a DML statement}

\item{queryPlan}{[Output-only, Experimental] Describes execution plan for the query}

\item{referencedTables}{[Output-only, Experimental] Referenced tables for the job}

\item{schema}{[Output-only, Experimental] The schema of the results}

\item{statementType}{[Output-only, Experimental] The type of query statement, if valid}

\item{totalBytesBilled}{[Output-only] Total bytes billed for the job}

\item{totalBytesProcessed}{[Output-only] Total bytes processed for the job}

\item{undeclaredQueryParameters}{[Output-only, Experimental] Standard SQL only: list of undeclared query parameters detected during a dry run validation}
}
\value{
JobStatistics2 object
}
\description{
JobStatistics2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
