% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobStatistics3}
\alias{JobStatistics3}
\title{JobStatistics3 Object}
\usage{
JobStatistics3(inputFileBytes = NULL, inputFiles = NULL,
  outputBytes = NULL, outputRows = NULL)
}
\arguments{
\item{inputFileBytes}{[Output-only] Number of bytes of source data in a load job}

\item{inputFiles}{[Output-only] Number of source files in a load job}

\item{outputBytes}{[Output-only] Size of the loaded data in bytes}

\item{outputRows}{[Output-only] Number of rows imported in a load job}
}
\value{
JobStatistics3 object
}
\description{
JobStatistics3 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

