% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobStatistics4}
\alias{JobStatistics4}
\title{JobStatistics4 Object}
\usage{
JobStatistics4(destinationUriFileCounts = NULL)
}
\arguments{
\item{destinationUriFileCounts}{[Output-only] Number of files per destination URI or URI pattern specified in the extract configuration}
}
\value{
JobStatistics4 object
}
\description{
JobStatistics4 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

