% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JobStatus}
\alias{JobStatus}
\title{JobStatus Object}
\usage{
JobStatus(errorResult = NULL, errors = NULL, state = NULL)
}
\arguments{
\item{errorResult}{[Output-only] Final error result of the job}

\item{errors}{[Output-only] All errors encountered during the running of the job}

\item{state}{[Output-only] Running state of the job}
}
\value{
JobStatus object
}
\description{
JobStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

