% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{JsonObject}
\alias{JsonObject}
\title{JsonObject Object}
\usage{
JsonObject()
}
\value{
JsonObject object
}
\description{
JsonObject Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single JSON object.
}

