% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ProjectList}
\alias{ProjectList}
\title{ProjectList Object}
\usage{
ProjectList(ProjectList.projects = NULL, etag = NULL,
  nextPageToken = NULL, projects = NULL, totalItems = NULL)
}
\arguments{
\item{ProjectList.projects}{The \link{ProjectList.projects} object or list of objects}

\item{etag}{A hash of the page of results}

\item{nextPageToken}{A token to request the next page of results}

\item{projects}{Projects to which you have at least READ access}

\item{totalItems}{The total number of projects in the list}
}
\value{
ProjectList object
}
\description{
ProjectList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProjectList functions: \code{\link{ProjectList.projects}}
}

