% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ProjectReference}
\alias{ProjectReference}
\title{ProjectReference Object}
\usage{
ProjectReference(projectId = NULL)
}
\arguments{
\item{projectId}{[Required] ID of the project}
}
\value{
ProjectReference object
}
\description{
ProjectReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
