% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{projects.list}
\alias{projects.list}
\title{Lists all projects to which you have been granted any project role.}
\usage{
projects.list(maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Page token, returned by a previous call, to request the next page of results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}
