% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryParameter}
\alias{QueryParameter}
\title{QueryParameter Object}
\usage{
QueryParameter(name = NULL, parameterType = NULL, parameterValue = NULL)
}
\arguments{
\item{name}{[Optional] If unset, this is a positional parameter}

\item{parameterType}{[Required] The type of this parameter}

\item{parameterValue}{[Required] The value of this parameter}
}
\value{
QueryParameter object
}
\description{
QueryParameter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

