% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryParameterType}
\alias{QueryParameterType}
\title{QueryParameterType Object}
\usage{
QueryParameterType(QueryParameterType.structTypes = NULL, arrayType = NULL,
  structTypes = NULL, type = NULL)
}
\arguments{
\item{QueryParameterType.structTypes}{The \link{QueryParameterType.structTypes} object or list of objects}

\item{arrayType}{[Optional] The type of the array's elements, if this is an array}

\item{structTypes}{[Optional] The types of the fields of this struct, in order, if this is a struct}

\item{type}{[Required] The top level type of this field}
}
\value{
QueryParameterType object
}
\description{
QueryParameterType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other QueryParameterType functions: \code{\link{QueryParameterType.structTypes}}
}

