% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryParameterType.structTypes}
\alias{QueryParameterType.structTypes}
\title{QueryParameterType.structTypes Object}
\usage{
QueryParameterType.structTypes()
}
\value{
QueryParameterType.structTypes object
}
\description{
QueryParameterType.structTypes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other QueryParameterType functions: \code{\link{QueryParameterType}}
}
