% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryParameterValue}
\alias{QueryParameterValue}
\title{QueryParameterValue Object}
\usage{
QueryParameterValue(arrayValues = NULL, structValues = NULL, value = NULL)
}
\arguments{
\item{arrayValues}{[Optional] The array values, if this is an array type}

\item{structValues}{[Optional] The struct field values, in order of the struct type's declaration}

\item{value}{[Optional] The value of this value, if a simple scalar type}
}
\value{
QueryParameterValue object
}
\description{
QueryParameterValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

