% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryParameterValue.structValues}
\alias{QueryParameterValue.structValues}
\title{QueryParameterValue.structValues Object}
\usage{
QueryParameterValue.structValues()
}
\value{
QueryParameterValue.structValues object
}
\description{
QueryParameterValue.structValues Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] The struct field values, in order of the struct type's declaration.
}
\seealso{
Other QueryParameterValue functions: \code{\link{QueryParameterValue}}
}
