% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{QueryRequest}
\alias{QueryRequest}
\title{QueryRequest Object}
\usage{
QueryRequest(defaultDataset = NULL, dryRun = NULL, maxResults = NULL,
  parameterMode = NULL, preserveNulls = NULL, query = NULL,
  queryParameters = NULL, timeoutMs = NULL, useLegacySql = NULL,
  useQueryCache = NULL)
}
\arguments{
\item{defaultDataset}{[Optional] Specifies the default datasetId and projectId to assume for any unqualified table names in the query}

\item{dryRun}{[Optional] If set to true, BigQuery doesn't run the job}

\item{maxResults}{[Optional] The maximum number of rows of data to return per page of results}

\item{parameterMode}{[Experimental] Standard SQL only}

\item{preserveNulls}{[Deprecated] This property is deprecated}

\item{query}{[Required] A query string, following the BigQuery query syntax, of the query to execute}

\item{queryParameters}{[Experimental] Query parameters for Standard SQL queries}

\item{timeoutMs}{[Optional] How long to wait for the query to complete, in milliseconds, before the request times out and returns}

\item{useLegacySql}{[Experimental] Specifies whether to use BigQuery's legacy SQL dialect for this query}

\item{useQueryCache}{[Optional] Whether to look for the result in the query cache}
}
\value{
QueryRequest object
}
\description{
QueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other QueryRequest functions: \code{\link{jobs.query}}
}

