% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{Streamingbuffer}
\alias{Streamingbuffer}
\title{Streamingbuffer Object}
\usage{
Streamingbuffer(estimatedBytes = NULL, estimatedRows = NULL,
  oldestEntryTime = NULL)
}
\arguments{
\item{estimatedBytes}{[Output-only] A lower-bound estimate of the number of bytes currently in the streaming buffer}

\item{estimatedRows}{[Output-only] A lower-bound estimate of the number of rows currently in the streaming buffer}

\item{oldestEntryTime}{[Output-only] Contains the timestamp of the oldest entry in the streaming buffer, in milliseconds since the epoch, if the streaming buffer is available}
}
\value{
Streamingbuffer object
}
\description{
Streamingbuffer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
