% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{Table}
\alias{Table}
\title{Table Object}
\usage{
Table(Table.labels = NULL, creationTime = NULL, description = NULL,
  etag = NULL, expirationTime = NULL, externalDataConfiguration = NULL,
  friendlyName = NULL, id = NULL, labels = NULL,
  lastModifiedTime = NULL, location = NULL, numBytes = NULL,
  numLongTermBytes = NULL, numRows = NULL, schema = NULL,
  selfLink = NULL, streamingBuffer = NULL, tableReference = NULL,
  timePartitioning = NULL, type = NULL, view = NULL)
}
\arguments{
\item{Table.labels}{The \link{Table.labels} object or list of objects}

\item{creationTime}{[Output-only] The time when this table was created, in milliseconds since the epoch}

\item{description}{[Optional] A user-friendly description of this table}

\item{etag}{[Output-only] A hash of this resource}

\item{expirationTime}{[Optional] The time when this table expires, in milliseconds since the epoch}

\item{externalDataConfiguration}{[Optional] Describes the data format, location, and other properties of a table stored outside of BigQuery}

\item{friendlyName}{[Optional] A descriptive name for this table}

\item{id}{[Output-only] An opaque ID uniquely identifying the table}

\item{labels}{[Experimental] The labels associated with this table}

\item{lastModifiedTime}{[Output-only] The time when this table was last modified, in milliseconds since the epoch}

\item{location}{[Output-only] The geographic location where the table resides}

\item{numBytes}{[Output-only] The size of this table in bytes, excluding any data in the streaming buffer}

\item{numLongTermBytes}{[Output-only] The number of bytes in the table that are considered 'long-term storage'}

\item{numRows}{[Output-only] The number of rows of data in this table, excluding any data in the streaming buffer}

\item{schema}{[Optional] Describes the schema of this table}

\item{selfLink}{[Output-only] A URL that can be used to access this resource again}

\item{streamingBuffer}{[Output-only] Contains information regarding this table's streaming buffer, if one is present}

\item{tableReference}{[Required] Reference describing the ID of this table}

\item{timePartitioning}{[Experimental] If specified, configures time-based partitioning for this table}

\item{type}{[Output-only] Describes the table type}

\item{view}{[Optional] The view definition}
}
\value{
Table object
}
\description{
Table Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Table functions: \code{\link{Table.labels}},
  \code{\link{tables.insert}}, \code{\link{tables.patch}},
  \code{\link{tables.update}}
}
