% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableCell}
\alias{TableCell}
\title{TableCell Object}
\usage{
TableCell(v = NULL)
}
\arguments{
\item{v}{No description}
}
\value{
TableCell object
}
\description{
TableCell Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

