% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{tabledata.list}
\alias{tabledata.list}
\title{Retrieves table data from a specified set of rows. Requires the READER dataset role.}
\usage{
tabledata.list(projectId, datasetId, tableId, maxResults = NULL,
  pageToken = NULL, startIndex = NULL)
}
\arguments{
\item{projectId}{Project ID of the table to read}

\item{datasetId}{Dataset ID of the table to read}

\item{tableId}{Table ID of the table to read}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Page token, returned by a previous call, identifying the result set}

\item{startIndex}{Zero-based index of the starting row to read}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}
