% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableDataList}
\alias{TableDataList}
\title{TableDataList Object}
\usage{
TableDataList(etag = NULL, pageToken = NULL, rows = NULL,
  totalRows = NULL)
}
\arguments{
\item{etag}{A hash of this page of results}

\item{pageToken}{A token used for paging results}

\item{rows}{Rows of results}

\item{totalRows}{The total number of rows in the complete table}
}
\value{
TableDataList object
}
\description{
TableDataList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

