% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableFieldSchema}
\alias{TableFieldSchema}
\title{TableFieldSchema Object}
\usage{
TableFieldSchema(description = NULL, fields = NULL, mode = NULL,
  name = NULL, type = NULL)
}
\arguments{
\item{description}{[Optional] The field description}

\item{fields}{[Optional] Describes the nested schema fields if the type property is set to RECORD}

\item{mode}{[Optional] The field mode}

\item{name}{[Required] The field name}

\item{type}{[Required] The field data type}
}
\value{
TableFieldSchema object
}
\description{
TableFieldSchema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

