% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableList}
\alias{TableList}
\title{TableList Object}
\usage{
TableList(TableList.tables = NULL, etag = NULL, nextPageToken = NULL,
  tables = NULL, totalItems = NULL)
}
\arguments{
\item{TableList.tables}{The \link{TableList.tables} object or list of objects}

\item{etag}{A hash of this page of results}

\item{nextPageToken}{A token to request the next page of results}

\item{tables}{Tables in the requested dataset}

\item{totalItems}{The total number of tables in the dataset}
}
\value{
TableList object
}
\description{
TableList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TableList functions: \code{\link{TableList.tables}}
}

