% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableList.tables.view}
\alias{TableList.tables.view}
\title{TableList.tables.view Object}
\usage{
TableList.tables.view(useLegacySql = NULL)
}
\arguments{
\item{useLegacySql}{True if view is defined in legacy SQL dialect, false if in standard SQL}
}
\value{
TableList.tables.view object
}
\description{
TableList.tables.view Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional details for a view.
}
\seealso{
Other TableList functions: \code{\link{TableList.tables.labels}},
  \code{\link{TableList.tables}}, \code{\link{TableList}}
}
