% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableReference}
\alias{TableReference}
\title{TableReference Object}
\usage{
TableReference(datasetId = NULL, projectId = NULL, tableId = NULL)
}
\arguments{
\item{datasetId}{[Required] The ID of the dataset containing this table}

\item{projectId}{[Required] The ID of the project containing this table}

\item{tableId}{[Required] The ID of the table}
}
\value{
TableReference object
}
\description{
TableReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

