% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableRow}
\alias{TableRow}
\title{TableRow Object}
\usage{
TableRow(f = NULL)
}
\arguments{
\item{f}{Represents a single row in the result set, consisting of one or more fields}
}
\value{
TableRow object
}
\description{
TableRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
