% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{tables.delete}
\alias{tables.delete}
\title{Deletes the table specified by tableId from the dataset. If the table contains data, all the data will be deleted.}
\usage{
tables.delete(projectId, datasetId, tableId)
}
\arguments{
\item{projectId}{Project ID of the table to delete}

\item{datasetId}{Dataset ID of the table to delete}

\item{tableId}{Table ID of the table to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}

