% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{tables.get}
\alias{tables.get}
\title{Gets the specified table resource by table ID. This method does not return the data in the table, it only returns the table resource, which describes the structure of this table.}
\usage{
tables.get(projectId, datasetId, tableId)
}
\arguments{
\item{projectId}{Project ID of the requested table}

\item{datasetId}{Dataset ID of the requested table}

\item{tableId}{Table ID of the requested table}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}
}
