% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_functions.R
\name{tables.insert}
\alias{tables.insert}
\title{Creates a new, empty table in the dataset.}
\usage{
tables.insert(Table, projectId, datasetId)
}
\arguments{
\item{Table}{The \link{Table} object to pass to this method}

\item{projectId}{Project ID of the new table}

\item{datasetId}{Dataset ID of the new table}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/bigquery
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/bigquery, https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/bigquery/}{Google Documentation}

Other Table functions: \code{\link{Table.labels}},
  \code{\link{Table}}, \code{\link{tables.patch}},
  \code{\link{tables.update}}
}
