% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TableSchema}
\alias{TableSchema}
\title{TableSchema Object}
\usage{
TableSchema(fields = NULL)
}
\arguments{
\item{fields}{Describes the fields in a table}
}
\value{
TableSchema object
}
\description{
TableSchema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

