% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{TimePartitioning}
\alias{TimePartitioning}
\title{TimePartitioning Object}
\usage{
TimePartitioning(expirationMs = NULL, type = NULL)
}
\arguments{
\item{expirationMs}{[Optional] Number of milliseconds for which to keep the storage for a partition}

\item{type}{[Required] The only type supported is DAY, which will generate one partition per day based on data loading time}
}
\value{
TimePartitioning object
}
\description{
TimePartitioning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
