% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{UserDefinedFunctionResource}
\alias{UserDefinedFunctionResource}
\title{UserDefinedFunctionResource Object}
\usage{
UserDefinedFunctionResource(inlineCode = NULL, resourceUri = NULL)
}
\arguments{
\item{inlineCode}{[Pick one] An inline resource that contains code for a user-defined function (UDF)}

\item{resourceUri}{[Pick one] A code resource to load from a Google Cloud Storage URI (gs://bucket/path)}
}
\value{
UserDefinedFunctionResource object
}
\description{
UserDefinedFunctionResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
