% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{ViewDefinition}
\alias{ViewDefinition}
\title{ViewDefinition Object}
\usage{
ViewDefinition(query = NULL, useLegacySql = NULL,
  userDefinedFunctionResources = NULL)
}
\arguments{
\item{query}{[Required] A query that BigQuery executes when the view is referenced}

\item{useLegacySql}{[Experimental] Specifies whether to use BigQuery's legacy SQL for this view}

\item{userDefinedFunctionResources}{[Experimental] Describes user-defined function resources used in the query}
}
\value{
ViewDefinition object
}
\description{
ViewDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

