% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery_objects.R
\name{Dataset.labels}
\alias{Dataset.labels}
\title{Dataset.labels Object}
\usage{
Dataset.labels()
}
\value{
Dataset.labels object
}
\description{
Dataset.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Experimental] The labels associated with this dataset. You can use these to organize and group your datasets. You can set this property when inserting or updating a dataset. Label keys and values can be no longer than 63 characters, can only contain letters, numeric characters, underscores and dashes. International characters are allowed. Label values are optional. Label keys must start with a letter and must be unique within a dataset. Both keys and values are additionally constrained to be <= 128 bytes in size.
}
\seealso{
Other Dataset functions: \code{\link{Dataset.access}},
  \code{\link{Dataset}}, \code{\link{datasets.insert}},
  \code{\link{datasets.patch}},
  \code{\link{datasets.update}}
}

